Move logic to utils file
