+++ src/components/feature-board/shared/FileInfoChips.vue
@@ -8,6 +8,10 @@
 <script lang="ts">
 import { computed, defineComponent, PropType } from "vue";
 import { FileRecordResponse } from "@/fileStorageApi/v3";
+import {
+	convertFileSizeToHumanReadable,
+	getFileExtension,
+} from "@/utils/fileHelper";
 
 export default defineComponent({
 	name: "FileInfoChips",
@@ -19,32 +23,13 @@ export default defineComponent({
 	},
 
 	setup(props) {
-		const fileSize = computed(() => {
-			const units = ["B", "KB", "MB", "GB", "TB"];
-			const threshold = 1024;
+		const fileSize = computed(() =>
+			convertFileSizeToHumanReadable(props.fileRecord.size)
+		);
 
-			let convertedSize = props.fileRecord.size;
-			let unit = units[0];
-			let power = 1;
-
-			while (convertedSize >= threshold && units.length >= power) {
-				convertedSize /= threshold;
-				unit = units[power];
-				power++;
-			}
-
-			const humanReadableFileSize = convertedSize.toFixed(2) + " " + unit;
-
-			return humanReadableFileSize;
-		});
-
-		const fileExtension = computed(() => {
-			const ext = props.fileRecord.name
-				.substring(props.fileRecord.name.lastIndexOf(".") + 1)
-				.toUpperCase();
-
-			return ext;
-		});
+		const fileExtension = computed(() =>
+			getFileExtension(props.fileRecord.name)
+		);
 
 		return {
 			fileSize,
